Unicode character support and string cast
I was having some problems with movies that had unicode characters such as "⅓". They were stopping the script dead in its tracks. I'm not the best at python but adding utf8 encoding seemed to fix the problems I was having. 

Also movies that were only numbers, such as "300" were failing to be written to the file. So I cast everything to a string before it was written.